refactor: api/users/:id GET, POST, PUT
