+++ README.md
@@ -1,1 +1,15 @@
 # node-crud-api
+
+[Assignment](https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/crud-api/assignment.md)
+
+[Study materials](https://github.com/rolling-scopes-school/tasks/blob/master/node/modules/crud-api/README.md)
+
+# TODO
+
+<!-- The repository with the application contains a Readme.md file containing detailed instructions for installing, running and using the application -->
+
+# Installing
+
+# Running
+
+# Using
