+++ src/users/user.ts
@@ -1,0 +1,8 @@
+export interface User {
+  id?: UserId;
+  username: string;
+  age: number;
+  hobbies: string[];
+}
+
+export type UserId = string;
